Adding tests, minor fixes to work on tutorial notebooks
