+++ astrophot/models/__init__.py
@@ -24,3 +24,4 @@ from .nuker_model import *
 from .zernike_model import *
 from .airy_psf import *
 from .point_source import *
+from .group_psf_model import *
